Разделяет код проекта на модули
