+++ js/data.js
@@ -1,0 +1,13 @@
+const title = ['Уютная квартирка в центре Токио', 'Квартира в Москве', 'Дом с небольшим двориком', 'Комната в общежитии',];
+const  type = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const checkin = ['12:00', '13:00', '14:00'];
+const checkout = ['12:00', '13:00', '14:00'];
+const features = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const description = ['Большое светлое пространство', 'Минималистичный дизайн', 'Свежий ремонт'];
+const photos = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+export {title, type, checkin, checkout, features, description, photos};
